+++ src/main/bookmark-store.ts
@@ -173,6 +173,10 @@ export class BookmarkStore {
     })
   }
 
+  flush(): Promise<void> {
+    return this.mutationQueue
+  }
+
   private queueMutation<T>(mutation: () => Promise<T>): Promise<T> {
     const operation = this.mutationQueue.then(mutation)
     this.mutationQueue = operation.then(() => undefined, () => undefined)