+++ src/main/commercial-license-store.ts
@@ -164,6 +164,10 @@ export class CommercialLicenseStore {
     }
   }
 
+  flush(): Promise<void> {
+    return this.mutationQueue
+  }
+
   private queueMutation<T>(mutation: () => Promise<T>): Promise<T> {
     const operation = this.mutationQueue.then(mutation)
     this.mutationQueue = operation.then(() => undefined, () => undefined)