+++ src/main/site-permission-store.ts
@@ -124,6 +124,10 @@ export class SitePermissionStore {
     })
   }
 
+  flush(): Promise<void> {
+    return this.mutationQueue
+  }
+
   private queueMutation<T>(mutation: () => Promise<T>): Promise<T> {
     const operation = this.mutationQueue.then(mutation)
     this.mutationQueue = operation.then(() => undefined, () => undefined)